noteai/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "noteai"
version = "1.0.0"
description = "AI-powered note-taking and sharing platform"
authors = [
    {name = "NoteAI Team", email = "dev@noteai.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Web框架
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    
    # AutoGen AI框架
    "autogen-agentchat>=0.4.0",
    "autogen-ext[openai]>=0.4.0", 
    "autogen-core>=0.4.0",
    
    # 数据库
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "motor>=3.3.0",
    "pymongo>=4.6.0",
    "redis>=5.0.0",
    
    # 认证和安全
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    
    # 数据验证
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    
    # 任务队列
    "celery>=5.3.0",
    "kombu>=5.3.0",
    
    # 搜索
    "elasticsearch>=8.11.0",
    
    # 工具库
    "httpx>=0.25.0",
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
    "markdown>=3.5.0",
    "Pillow>=10.1.0",
]

[project.optional-dependencies]
dev = [
    # 测试
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.25.0",
    "factory-boy>=3.3.0",
    "faker>=20.1.0",
    
    # 代码质量
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
    
    # 开发工具
    "ipython>=8.17.0",
    "jupyter>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/noteai/noteai"
Repository = "https://github.com/noteai/noteai.git"
Documentation = "https://docs.noteai.com"

[tool.setuptools.packages.find]
where = ["."]
include = ["noteai*"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["noteai"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "motor.*",
    "pymongo.*",
    "celery.*",
    "elasticsearch.*",
    "autogen_agentchat.*",
    "autogen_ext.*",
    "autogen_core.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
